Move image parsing code into the ImageResponseParser
